Added spec-purpose handlers (file, customConsole)
Besides the general-porpuse custom handlers, I added support for writable streams (like files for example) so that msgs output there too.

I also added custom console handler support, for the cases when one wants to altter the message being outputted to the console itself, or if it's an empty function, then one can hide console output (eg hide debug msgs (but still write them to file) for the times when not debugging.)

Also, some inner functions has been exposed to all the handlers through fn.call() so that one can colorize or getStringContent of the msg if wishes.